Don't redownload MAME archive

+++ runners/mame/build.sh
@@ -23,7 +23,11 @@ release=$(curl http://mamedev.org/release.html | grep -E "href.*s.zip" | cut -d"
 version=$(curl http://mamedev.org/release.html | grep -E -o "release is version [\.0-9]+" | grep -E -o 0.[0-9]+)
 archive=$(echo ${release} | cut -d"/" -f 9)
 
-wget "${release}" -O ${archive}
+if [ -f "$archive" ]; then
+    echo "$achive already available"
+else
+    wget "${release}" -O ${archive}
+fi
 unzip -o $archive
 
 mkdir -p ${source_dir}
